Added pre-upgrade check to masters section
I think, I today ran into the problem that I upgraded a cluster that was not stable in the beginning. To avoid other users run into the same I added a little pre-upgrade check to the masters section. 

I am not sure if this is true before the upgrade: 
"Verify that curl http://<dcos_master_private_ip>:5050/metrics/snapshot has the metric registrar/log/recovered with a value of 1 for each master."